refactor: make UIBaseObject implements BaseObject
